requirements.txt
typer[all]
openai
notion-client
atlassian-python-api
# faiss-cpu -- to handle late
chromadb
pyyaml
